+++ packages/app-project/src/screens/ProjectAboutPage/ProjectAboutPageConnector.js
@@ -1,34 +1,29 @@
 import { observer, MobXProviderContext } from 'mobx-react'
-import React, { useContext } from 'react'
+import React from 'react'
 import { arrayOf, bool, object, shape, string } from 'prop-types'
 
 import ProjectAboutPage from './ProjectAboutPage'
 
 /**
-  This is for enzyme tests.
-*/
-function useStoreContext (testStore) {
-  const { store } = testStore || useContext(MobXProviderContext)
-  return {
-    store
-  }
-}
-
-/**
   Connect the about page to the store. Pass down correct aboutPages data.
 */
-function ProjectAboutPageConnector ({ pageType, teamArray, testStore }) {
-  const { store } = useStoreContext(testStore)
-  const { inBeta } = store.project
-  const aboutPages = store.project.about_pages
-  const projectDisplayName = store.project.display_name
-  const [aboutPageData] = aboutPages.filter(page => page.url_key === pageType)
+function ProjectAboutPageConnector ({ pageType, teamArray }) {
+  const {
+    store: {
+      project: {
+        inBeta = false,
+        about_pages = [],
+        display_name = ''
+      }
+    }
+  } = React.useContext(MobXProviderContext)
+  const [aboutPageData] = about_pages.filter(page => page.url_key === pageType)
   return aboutPageData ? (
     <ProjectAboutPage
       inBeta={inBeta}
       aboutPageData={aboutPageData}
       teamArray={teamArray}
-      projectDisplayName={projectDisplayName}
+      projectDisplayName={display_name}
     />
   ) : (
     <p>No data for this page...</p>
@@ -46,19 +41,7 @@ ProjectAboutPageConnector.propTypes = {
       login: string,
       role: string
     })
-  ),
-  /**
-   This store is used in enzyme tests
-  */
-  testStore: shape({
-    store: shape({
-      project: shape({
-        about_pages: arrayOf(shape({
-          id: string.isRequired
-        }))
-      })
-    })
-  })
+  )
 }
 
 export default observer(ProjectAboutPageConnector)