+++ packages/app-project/src/screens/ProjectAboutPage/ProjectAboutPageConnector.spec.js
@@ -1,21 +1,27 @@
 import { shallow } from 'enzyme'
 import React from 'react'
+import sinon from 'sinon'
 
 import ProjectAboutPageConnector from './ProjectAboutPageConnector'
 import ProjectAboutPage from './ProjectAboutPage'
 
 describe('Component > ProjectAboutPageConnector', function () {
   let wrapper
-  const testStore = {
-    store: {
-      project: {
-        about_pages: [ { id: '1234', title: 'Title', url_key: 'science_case' } ],
-        inBeta: false
-      }
-    }
-  }
+  let useContextStub
   before(function () {
-    wrapper = shallow(<ProjectAboutPageConnector testStore={testStore} pageType='science_case' />)
+    useContextStub = sinon.stub(React, 'useContext').callsFake(() => {
+      return {
+        store: {
+          project: {
+            about_pages: [
+              { id: '1234', title: 'Title', url_key: 'science_case' }
+            ],
+            inBeta: false
+          }
+        }
+      }
+    })
+    wrapper = shallow(<ProjectAboutPageConnector pageType='science_case' />)
   })
 
   it('should render without crashing', function () {
@@ -26,5 +32,9 @@ describe('Component > ProjectAboutPageConnector', function () {
     const aboutPage = wrapper.find(ProjectAboutPage)
     const aboutProps = aboutPage.prop('aboutPageData')
     expect(aboutProps.url_key).to.equal('science_case')
+  })
+
+  after(function () {
+    useContextStub.restore()
   })
 })