+++ packages/lib-classifier/src/plugins/drawingTools/experimental/components/TranscriptionLine/TranscriptionLine.stories.js
@@ -70,7 +70,10 @@ function setupStores() {
       active: { id: cuid() }
     },
     workflowSteps: {
-      activeStepTasks: [drawingTask]
+      activeInteractionTask: {},
+      activeStepTasks: [drawingTask],
+      findTasksByType: () => { return [] },
+      interactionTask: {}
     }
   }